+++ python/kserve/kserve/protocol/rest/openai/endpoints.py
@@ -178,4 +178,4 @@ def register_openai_endpoints(app: FastAPI, dataplane: OpenAIDataPlane):
         r"/v1/models/{model_name}", endpoints.health, methods=["GET"]
     )
     app.include_router(openai_router)
-    app.add_exception_handler(OpenAIError, openai_error_handler)
+    app.add_exception_handler(OpenAIError, openai_error_handler)  # TODO: double check
